delete record.py
